Wave 9: 129 functions matched, 631 -> 760, and the m4a split caught in the act
Ten agents over 133 scheduled functions: 129 matched, 1 reclassified as
hand-written, 3 parked. Both builds reproduce SHA1 14dd0b22. promote.py
produced compiling C first time across all 129, which is the fix from
e52bd7d holding.

The brief told agents that an FE-named function inside the m4a span is
almost certainly the m4a.c side. That heuristic is wrong and an agent
overruled it on evidence. sub_08070300 is FE-named clear_modM and takes the
track pointer in r1 with r0 never read; both call sites confirm it, and the
COMPILED copy of the same routine exists separately at sub_08071564 with the
two zero stores in the opposite order. One C source cannot produce both.
The calling convention outranks the FE name, and asm-resident.json now says
so in its header rather than repeating the disproved blanket claim about the
span. 25 excluded.

Three parked, all size-exact or near it, all with the mechanism identified
and the dead ends written down so no later wave re-derives them:

  sub_0801A6C0  98.4%, ONE byte. An allocator tie-break: which of the two
                values live across the loop, the base pointer or the hoisted
                zero, gets r3. 106,936 permuter iterations found nothing
                better. Settled on the way: the function is non-void with no
                return statement, and naming a global both inside and after
                a loop triggers the -fforce-addr .rodata indirection unless
                &g is bound to a local first.
  sub_080713F8  MPlayTempoControl, 32 of 40 bytes. The ROM holds one more
                value live across the multiply, which forces the r4
                allocation and the frame. The wave-8 free-liveness idiom
                does not reach it: that only vanishes on a WORD member, and
                the halfword form deletes the load along with the liveness.
  sub_08071564  MP_clear_modM, 28 of 32 bytes. The ROM reads track->flags
                after the conditional merges; every `?:` spelling hoists it
                above the cmp and every if/else statement form collapses to
                one arm. Eighteen spellings in one probe, now a worked
                example in the Control flow section.

sub_0801F838 (fan_in 16, the highest in the queue) and sub_080094EC
(fan_in 12) both landed, so two large caller sets are now type-constrained.

+++ include/unknown-functions.h
@@ -33,8 +33,15 @@ void sub_08012C58(void *); // possibly "SetupBackgrounds"
 void sub_08011B34(void *);
 void sub_08011B5C(void *);
 
-void sub_08013098(s16, ProcPtr);
-void sub_080130B0(s16, ProcPtr);
+/* `int`, not `s16` -- the same correction wave 7 made to the eight starters
+ * above, and for the same reason. Both open with a bare `adds r4, r0, #0`;
+ * declaring the parameter `s16` adds `lsl #16; lsr #16` (PROMOTE_MODE widening
+ * a narrow parameter that has to survive the `bl`) and cannot match. Their only
+ * call sites are ProcCmd_1D_0801D0E4/ProcCmd_1E_0801D104 in proc.c, which pass
+ * `dataImm` -- already sign-extended by its `ldrsh`, so the caller side emits
+ * the same code either way. */
+void sub_08013098(int, ProcPtr);
+void sub_080130B0(int, ProcPtr);
 /* Starts gUnknown_0848936C and stashes three halfwords at +0x64/+0x66/+0x68 of
  * the new proc; a NULL fourth argument selects Proc_Start on tree 3 over
  * Proc_StartBlocking. The three values are bare `strh` stores, which fixes them
